Split Helpers and Extensions
