+++ src/Bannerlord.ButterLib.Implementation/Common/Extensions/CampaignExtensionsImplementation.cs
@@ -1,6 +1,6 @@
 ﻿using Bannerlord.ButterLib.CampaignIdentifier;
-using Bannerlord.ButterLib.Common.Extensions;
 using Bannerlord.ButterLib.DistanceMatrix;
+using Bannerlord.ButterLib.Extensions;
 using Bannerlord.ButterLib.Implementation.CampaignIdentifier.CampaignBehaviors;
 using Bannerlord.ButterLib.Implementation.DistanceMatrix;
