+++ src/Bannerlord.ButterLib/Extensions/CampaignExtensions.cs
@@ -1,16 +1,33 @@
 ﻿using Bannerlord.ButterLib.CampaignIdentifier;
 using Bannerlord.ButterLib.DistanceMatrix;
+using Bannerlord.ButterLib.Extensions;
+
+using Microsoft.Extensions.DependencyInjection;
 
 using TaleWorlds.CampaignSystem;
 
+// ReSharper disable once CheckNamespace
 namespace Bannerlord.ButterLib.Common.Extensions
 {
-    internal interface ICampaignExtensions
+    /// <summary>
+    /// Helper extension of the <see cref="Campaign" /> class
+    /// returning additional information, provided by the ButterLib.
+    /// </summary>
+    /// <remarks>
+    /// Contains easy accessible getters for the current CampaignId and <see cref="CampaignDescriptor" />
+    /// provided by CampaignIdentifier service as well as various geopolitical distance matrices
+    /// held in <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" /> objects.
+    /// </remarks>
+    public static class CampaignExtensions
     {
+        private static ICampaignExtensions? _instance;
+        private static ICampaignExtensions? Instance =>
+            _instance ??= ButterLibSubModule.Instance?.GetServiceProvider()?.GetRequiredService<ICampaignExtensions>();
+
         /// <summary>Gets ID of the <see cref="Campaign" />.</summary>
         /// <param name="campaign">The campaign</param>
         /// <returns>Alphanumeric string key for the campaign or null if campaign is not started or not yet identified.</returns>
-        string? GetCampaignId(Campaign campaign);
+        public static string? GetCampaignId(this Campaign campaign) => Instance?.GetCampaignId(campaign);
 
         /// <summary>Gets CampaignDescriptor for the <see cref="Campaign" />.</summary>
         /// <param name="campaign">The campaign</param>
@@ -18,7 +35,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// <see cref="CampaignDescriptor" /> object corresponding with the campaign
         /// or null if campaign is not started or not yet identified.
         /// </returns>
-        CampaignDescriptor? GetCampaignDescriptor(Campaign campaign);
+        public static CampaignDescriptor? GetCampaignDescriptor(this Campaign campaign) => Instance?.GetCampaignDescriptor(campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -29,7 +46,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// Distance matrix for all the towns, castles and villages of the current campaign,
         /// or null if the campaign has not started yet.
         /// </returns>
-        DistanceMatrix<Settlement>? GetDefaultSettlementDistanceMatrix(Campaign campaign);
+        public static DistanceMatrix<Settlement>? GetDefaultSettlementDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultSettlementDistanceMatrix(campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -41,7 +58,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// or null if the campaign has not started yet.
         /// </returns>
         /// <remarks>Calculation is based on the average distance between clans fiefs weighted by the fief type.</remarks>
-        DistanceMatrix<Clan>? GetDefaultClanDistanceMatrix(Campaign campaign);
+        public static DistanceMatrix<Clan>? GetDefaultClanDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultClanDistanceMatrix(campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -53,6 +70,6 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// or null if the campaign has not started yet.
         /// </returns>
         /// <remarks>Calculation is based on the average distance between kingdoms fiefs weighted by the fief type.</remarks>
-        DistanceMatrix<Kingdom>? GetDefaultKingdomDistanceMatrix(Campaign campaign);
+        public static DistanceMatrix<Kingdom>? GetDefaultKingdomDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultKingdomDistanceMatrix(campaign);
     }
 }
