+++ src/Bannerlord.ButterLib/Extensions/DictionaryExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 
+// ReSharper disable once CheckNamespace
 namespace Bannerlord.ButterLib.Common.Extensions
 {
     /// <summary>Extension class for working with Dictionaries.</summary>
