+++ src/Bannerlord.ButterLib/Extensions/ICampaignExtensions.cs
@@ -1,31 +1,17 @@
 ﻿using Bannerlord.ButterLib.CampaignIdentifier;
 using Bannerlord.ButterLib.DistanceMatrix;
 
-using Microsoft.Extensions.DependencyInjection;
-
 using TaleWorlds.CampaignSystem;
 
-namespace Bannerlord.ButterLib.Common.Extensions
-{
-    /// <summary>
-    /// Helper extension of the <see cref="Campaign" /> class
-    /// returning additional information, provided by the ButterLib.
-    /// </summary>
-    /// <remarks>
-    /// Contains easy accessible getters for the current CampaignId and <see cref="CampaignDescriptor" />
-    /// provided by CampaignIdentifier service as well as various geopolitical distance matrices
-    /// held in <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" /> objects.
-    /// </remarks>
-    public static class CampaignExtensions
-    {
-        private static ICampaignExtensions? _instance;
-        private static ICampaignExtensions? Instance =>
-            _instance ??= ButterLibSubModule.Instance?.GetServiceProvider()?.GetRequiredService<ICampaignExtensions>();
 
+namespace Bannerlord.ButterLib.Extensions
+{
+    internal interface ICampaignExtensions
+    {
         /// <summary>Gets ID of the <see cref="Campaign" />.</summary>
         /// <param name="campaign">The campaign</param>
         /// <returns>Alphanumeric string key for the campaign or null if campaign is not started or not yet identified.</returns>
-        public static string? GetCampaignId(this Campaign campaign) => Instance?.GetCampaignId(campaign);
+        string? GetCampaignId(Campaign campaign);
 
         /// <summary>Gets CampaignDescriptor for the <see cref="Campaign" />.</summary>
         /// <param name="campaign">The campaign</param>
@@ -33,7 +19,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// <see cref="CampaignDescriptor" /> object corresponding with the campaign
         /// or null if campaign is not started or not yet identified.
         /// </returns>
-        public static CampaignDescriptor? GetCampaignDescriptor(this Campaign campaign) => Instance?.GetCampaignDescriptor(campaign);
+        CampaignDescriptor? GetCampaignDescriptor(Campaign campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -44,7 +30,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// Distance matrix for all the towns, castles and villages of the current campaign,
         /// or null if the campaign has not started yet.
         /// </returns>
-        public static DistanceMatrix<Settlement>? GetDefaultSettlementDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultSettlementDistanceMatrix(campaign);
+        DistanceMatrix<Settlement>? GetDefaultSettlementDistanceMatrix(Campaign campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -56,7 +42,7 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// or null if the campaign has not started yet.
         /// </returns>
         /// <remarks>Calculation is based on the average distance between clans fiefs weighted by the fief type.</remarks>
-        public static DistanceMatrix<Clan>? GetDefaultClanDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultClanDistanceMatrix(campaign);
+        DistanceMatrix<Clan>? GetDefaultClanDistanceMatrix(Campaign campaign);
 
         /// <summary>
         /// Gets the <see cref="T:Bannerlord.ButterLib.DistanceMatrix.DistanceMatrix`1" />
@@ -68,6 +54,6 @@ namespace Bannerlord.ButterLib.Common.Extensions
         /// or null if the campaign has not started yet.
         /// </returns>
         /// <remarks>Calculation is based on the average distance between kingdoms fiefs weighted by the fief type.</remarks>
-        public static DistanceMatrix<Kingdom>? GetDefaultKingdomDistanceMatrix(this Campaign campaign) => Instance?.GetDefaultKingdomDistanceMatrix(campaign);
+        DistanceMatrix<Kingdom>? GetDefaultKingdomDistanceMatrix(Campaign campaign);
     }
 }
