+++ src/Bannerlord.ButterLib/Helpers/ApplicationVersionComparer.cs
@@ -2,6 +2,7 @@
 
 using TaleWorlds.Library;
 
+// ReSharper disable once CheckNamespace
 namespace Bannerlord.ButterLib.Common.Helpers
 {
     public class ApplicationVersionComparer : IComparer<ApplicationVersion>
