+++ src/Bannerlord.ButterLib/Helpers/ModuleInfo/LoadType.cs
@@ -1,0 +1,8 @@
+﻿namespace Bannerlord.ButterLib.Helpers.ModuleInfo
+{
+    public enum LoadType
+    {
+        LoadAfterThis,
+        LoadBeforeThis
+    }
+}
