+++ src/Bannerlord.ModuleLoader/Bannerlord.ModuleLoader.csproj
@@ -16,11 +16,13 @@
   <!--Module Creation-->
 
   <ItemGroup>
-    <Compile Include="..\Bannerlord.ButterLib\Common\Extensions\DictionaryExtensions.cs" Link="ButterLib\DictionaryExtensions.cs" />
-    <Compile Include="..\Bannerlord.ButterLib\Common\Helpers\AccessTools2.cs" Link="ButterLib\AccessTools2.cs" />
-    <Compile Include="..\Bannerlord.ButterLib\Common\Helpers\ExtendedModuleInfo.cs" Link="ButterLib\ExtendedModuleInfo.cs" />
-    <Compile Include="..\Bannerlord.ButterLib\Common\Helpers\ExtendedSubModuleInfo.cs" Link="ButterLib\ExtendedSubModuleInfo.cs" />
-    <Compile Include="..\Bannerlord.ButterLib\Common\Helpers\ModuleInfoHelper.cs" Link="ButterLib\ModuleInfoHelper.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Extensions\DictionaryExtensions.cs" Link="ButterLib\DictionaryExtensions.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\Harmony\AccessTools2.cs" Link="ButterLib\AccessTools2.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\ModuleInfo\ExtendedModuleInfo.cs" Link="ButterLib\ExtendedModuleInfo.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\ModuleInfo\LoadType.cs" Link="ButterLib\LoadType.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\ModuleInfo\DependedModuleMetadata.cs" Link="ButterLib\DependedModuleMetadata.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\ModuleInfo\ExtendedSubModuleInfo.cs" Link="ButterLib\ExtendedSubModuleInfo.cs" />
+    <Compile Include="..\Bannerlord.ButterLib\Helpers\ModuleInfo\ModuleInfoHelper.cs" Link="ButterLib\ModuleInfoHelper.cs" />
     <Compile Include="..\Bannerlord.ButterLib\Internal\NullableAttributes.cs" Link="ButterLib\NullableAttributes.cs" />
     <Compile Include="..\Bannerlord.ButterLib\SubModuleWrappers\MBSubModuleBaseWrapper.cs" Link="ButterLib\MBSubModuleBaseWrapper.cs" />
   </ItemGroup>
