+++ tests/Bannerlord.ButterLib.Implementation.Tests/DependencyInjectionTests.cs
@@ -3,6 +3,7 @@ using Bannerlord.ButterLib.Common.Extensions;
 using Bannerlord.ButterLib.Common.Helpers;
 using Bannerlord.ButterLib.DistanceMatrix;
 using Bannerlord.ButterLib.ExceptionHandler;
+using Bannerlord.ButterLib.Extensions;
 using Bannerlord.ButterLib.Implementation.CampaignIdentifier;
 using Bannerlord.ButterLib.Implementation.Common.Extensions;
 using Bannerlord.ButterLib.Implementation.DistanceMatrix;
